Cleanup part 1
