+++ engine/graphics/src/graphics.h
@@ -344,6 +344,33 @@ namespace dmGraphics
         uint32_t            m_StructSize;
     };
 
+    struct WriteAttributeParams
+    {
+        const VertexAttributeInfos*    m_VertexAttributeInfos;
+        const dmVMath::Matrix4*        m_WorldMatrix;
+        const dmVMath::Matrix4*        m_NormalMatrix;
+        const float*                   m_PositionsLocalSpace;
+        const float*                   m_PositionsWorldSpace;
+        const float*                   m_Normals;
+        const float*                   m_Tangents;
+        const float*                   m_Colors;
+        const float**                  m_UVChannels;
+        const float*                   m_PageIndices;
+        uint32_t                       m_PageIndicesCount;
+        uint32_t                       m_UVChannelsCount;
+
+        // Data source vector types corresponding to each data source passed in.
+        // The destination vector type is stored in the vertexattribute infos.
+        VertexAttribute::VectorType    m_PositionsVectorType;
+        VertexAttribute::VectorType    m_NormalsVectorType;
+        VertexAttribute::VectorType    m_TangentsVectorType;
+        VertexAttribute::VectorType    m_ColorsVectorType;
+        VertexAttribute::VectorType    m_PageIndicesVectorType;
+        VertexAttribute::VectorType    m_UVChannelsVectorType;
+
+        VertexStepFunction m_StepFunction;
+    };
+
     struct VertexDeclaration
     {
         struct Stream
@@ -582,37 +609,11 @@ namespace dmGraphics
     void             GetAttributeValues(const VertexAttribute& attribute, const uint8_t** data_ptr, uint32_t* data_size);
     Type             GetGraphicsType(VertexAttribute::DataType data_type);
 
-    struct WriteAttributeParams
-    {
-        const VertexAttributeInfos*    m_VertexAttributeInfos;
-        const dmVMath::Matrix4*        m_WorldMatrix;
-        const dmVMath::Matrix4*        m_NormalMatrix;
-        const float*                   m_PositionsLocalSpace;
-        const float*                   m_PositionsWorldSpace;
-        const float*                   m_Normals;
-        const float*                   m_Tangents;
-        const float*                   m_Colors;
-        const float**                  m_UVChannels;
-        const float*                   m_PageIndices;
-        uint32_t                       m_PageIndicesCount;
-        uint32_t                       m_UVChannelsCount;
+    float            VertexAttributeDataTypeToFloat(const dmGraphics::VertexAttribute::DataType data_type, const uint8_t* value_ptr);
+    uint8_t*         WriteVertexAttributeFromFloat(uint8_t* value_write_ptr, float value, dmGraphics::VertexAttribute::DataType data_type);
+    uint8_t*         WriteAttributes(uint8_t* write_ptr, uint32_t vertex_index, const WriteAttributeParams& params);
 
-        // Data source vector types corresponding to each data source passed in.
-        // The destination vector type is stored in the vertexattribute infos.
-        VertexAttribute::VectorType    m_PositionsVectorType;
-        VertexAttribute::VectorType    m_NormalsVectorType;
-        VertexAttribute::VectorType    m_TangentsVectorType;
-        VertexAttribute::VectorType    m_ColorsVectorType;
-        VertexAttribute::VectorType    m_PageIndicesVectorType;
-        VertexAttribute::VectorType    m_UVChannelsVectorType;
-
-        VertexStepFunction m_StepFunction;
-    };
-
-    float    VertexAttributeDataTypeToFloat(const dmGraphics::VertexAttribute::DataType data_type, const uint8_t* value_ptr);
-    uint8_t* WriteVertexAttributeFromFloat(uint8_t* value_write_ptr, float value, dmGraphics::VertexAttribute::DataType data_type);
-    uint8_t* WriteAttributes(uint8_t* write_ptr, uint32_t vertex_index, const WriteAttributeParams& params);
-
+    // Uniforms
     uint32_t         GetUniformName(HProgram prog, uint32_t index, char* buffer, uint32_t buffer_size, Type* type, int32_t* size);
     uint32_t         GetUniformCount(HProgram prog);
     HUniformLocation GetUniformLocation(HProgram prog, const char* name);
@@ -705,133 +706,6 @@ namespace dmGraphics
     // Asset handle helpers
     const char* GetAssetTypeLiteral(AssetType type);
     bool        IsAssetHandleValid(HContext context, HAssetHandle asset_handle);
-
-    static inline HAssetHandle MakeAssetHandle(HOpaqueHandle opaque_handle, AssetType asset_type)
-    {
-        assert(asset_type != ASSET_TYPE_NONE && "Invalid asset type");
-        uint64_t handle = ((uint64_t) asset_type) << 32 | opaque_handle;
-        assert(handle <= MAX_ASSET_HANDLE_VALUE);
-        return handle;
-    }
-
-    static inline AssetType GetAssetType(HAssetHandle asset_handle)
-    {
-        return (AssetType) (asset_handle >> 32);
-    }
-
-    static inline HOpaqueHandle GetOpaqueHandle(HAssetHandle asset_handle)
-    {
-        return (HOpaqueHandle) asset_handle & 0xFFFFFFFF;
-    }
-
-    static inline bool IsColorBufferType(BufferType buffer_type)
-    {
-        return buffer_type == BUFFER_TYPE_COLOR0_BIT ||
-               buffer_type == BUFFER_TYPE_COLOR1_BIT ||
-               buffer_type == BUFFER_TYPE_COLOR2_BIT ||
-               buffer_type == BUFFER_TYPE_COLOR3_BIT;
-    }
-
-    struct VertexAttributeInfoMetadata
-    {
-        uint32_t m_HasAttributeWorldPosition : 1;
-        uint32_t m_HasAttributeLocalPosition : 1;
-        uint32_t m_HasAttributeNormal        : 1;
-        uint32_t m_HasAttributeTangent       : 1;
-        uint32_t m_HasAttributeColor         : 1;
-        uint32_t m_HasAttributeTexCoord      : 1;
-        uint32_t m_HasAttributePageIndex     : 1;
-        uint32_t m_HasAttributeWorldMatrix   : 1;
-        uint32_t m_HasAttributeNormalMatrix  : 1;
-        uint32_t m_HasAttributeNone          : 1;
-    };
-
-    static inline VertexAttributeInfoMetadata GetVertexAttributeInfosMetaData(const VertexAttributeInfos& attribute_infos)
-    {
-        VertexAttributeInfoMetadata metadata = {};
-        for (int i = 0; i < attribute_infos.m_NumInfos; ++i)
-        {
-            const VertexAttributeInfo& info = attribute_infos.m_Infos[i];
-            switch(info.m_SemanticType)
-            {
-            case VertexAttribute::SEMANTIC_TYPE_POSITION:
-                metadata.m_HasAttributeWorldPosition |= info.m_CoordinateSpace == COORDINATE_SPACE_WORLD;
-                metadata.m_HasAttributeLocalPosition |= info.m_CoordinateSpace == COORDINATE_SPACE_LOCAL;
-                break;
-            case VertexAttribute::SEMANTIC_TYPE_TEXCOORD:
-                metadata.m_HasAttributeTexCoord = true;
-                break;
-            case VertexAttribute::SEMANTIC_TYPE_PAGE_INDEX:
-                metadata.m_HasAttributePageIndex = true;
-                break;
-            case VertexAttribute::SEMANTIC_TYPE_NORMAL:
-                metadata.m_HasAttributeNormal = true;
-                break;
-            case VertexAttribute::SEMANTIC_TYPE_TANGENT:
-                metadata.m_HasAttributeTangent = true;
-                break;
-            case VertexAttribute::SEMANTIC_TYPE_COLOR:
-                metadata.m_HasAttributeColor = true;
-                break;
-            case VertexAttribute::SEMANTIC_TYPE_WORLD_MATRIX:
-                metadata.m_HasAttributeWorldMatrix = true;
-                break;
-            case VertexAttribute::SEMANTIC_TYPE_NORMAL_MATRIX:
-                metadata.m_HasAttributeNormalMatrix = true;
-                break;
-            case VertexAttribute::SEMANTIC_TYPE_NONE:
-                metadata.m_HasAttributeNone = true;
-                break;
-            default:
-                break;
-            }
-        }
-        return metadata;
-    }
-
-    static inline bool IsTypeTextureType(Type type)
-    {
-        return type == TYPE_SAMPLER_2D ||
-               type == TYPE_SAMPLER_CUBE ||
-               type == TYPE_SAMPLER_2D_ARRAY ||
-               type == TYPE_IMAGE_2D;
-    }
-
-    static inline uint32_t GetLayerCount(TextureType type)
-    {
-        return type == TEXTURE_TYPE_CUBE_MAP ? 6 : 1;
-    }
-
-    static inline uint32_t VectorTypeToElementCount(VertexAttribute::VectorType vector_type)
-    {
-        switch(vector_type)
-        {
-            case VertexAttribute::VECTOR_TYPE_SCALAR: return 1;
-            case VertexAttribute::VECTOR_TYPE_VEC2:   return 2;
-            case VertexAttribute::VECTOR_TYPE_VEC3:   return 3;
-            case VertexAttribute::VECTOR_TYPE_VEC4:   return 4;
-            case VertexAttribute::VECTOR_TYPE_MAT2:   return 4;
-            case VertexAttribute::VECTOR_TYPE_MAT3:   return 9;
-            case VertexAttribute::VECTOR_TYPE_MAT4:   return 16;
-        }
-        return 0;
-    }
-
-    static inline uint32_t DataTypeToByteWidth(VertexAttribute::DataType data_type)
-    {
-        switch(data_type)
-        {
-            case dmGraphics::VertexAttribute::TYPE_BYTE:           return 1;
-            case dmGraphics::VertexAttribute::TYPE_UNSIGNED_BYTE:  return 1;
-            case dmGraphics::VertexAttribute::TYPE_SHORT:          return 2;
-            case dmGraphics::VertexAttribute::TYPE_UNSIGNED_SHORT: return 2;
-            case dmGraphics::VertexAttribute::TYPE_INT:            return 4;
-            case dmGraphics::VertexAttribute::TYPE_UNSIGNED_INT:   return 4;
-            case dmGraphics::VertexAttribute::TYPE_FLOAT:          return 4;
-            default: break;
-        }
-        return 0;
-    }
 
     /**
      * Get status of texture.
